+++ part_4/warehouse/src/main/java/vkaretko/products/Milk.java
@@ -18,7 +18,7 @@ public class Milk extends Food {
      * @param price price of product.
      * @param discount possible dicsount of product.
      */
-    public Milk(String name, Date expireDate, Date createDate, double price, int discount) {
+    public Milk(String name, Date expireDate, Date createDate, double price, double discount) {
         super(name, expireDate, createDate, price, discount);
     }
 }
